l-safe/compliance/src/utils.py
import os
import json
import logging
import time
from openai import OpenAI
from dotenv import load_dotenv

load_dotenv()

def setup_logger(name):
    logging.basicConfig(
        level=logging.INFO,
        format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
        datefmt='%H:%M:%S'
    )
    return logging.getLogger(name)

def get_client(base_url, api_key):
    return OpenAI(base_url=base_url, api_key=api_key)

def load_json(path):
    if not os.path.exists(path):
        raise FileNotFoundError(f"File not found: {path}")
    with open(path, 'r', encoding='utf-8') as f:
        return json.load(f)

def save_json(data, path):
    with open(path, 'w', encoding='utf-8') as f:
        json.dump(data, f, ensure_ascii=False, indent=2)