l-safe/compliance/requirements.txt
openai
tqdm
python-dotenv
argparse